+++ src/components/DriversCompanyDetailsForm/index.js
@@ -1,7 +1,9 @@
-import React, { useState } from 'react'
-import { useLanguage } from 'ordering-components-admin'
-import { DriversCompanyDetails as DriversCompanyDetailsController } from './naked'
-
+import React, { useState, useEffect } from 'react'
+import {
+  useLanguage,
+  DriversCompanyDetails as DriversCompanyDetailsController
+} from 'ordering-components-admin'
+import { Alert } from '../Confirm'
 import { DragScroll } from '../DragScroll'
 import { DriversCompanyGeneralDetails } from '../DriversCompanyGeneralDetails'
 import { DriversCompanyScheduleDetails } from '../DriversCompanyScheduleDetails'
@@ -17,11 +19,14 @@ import {
 const DriversCompanyDetailsFormUI = (props) => {
   const {
     driversCompany,
-    changesState
+    changesState,
+    actionState,
+    cleanActionState
   } = props
 
   const [, t] = useLanguage()
   const [currentTabItem, setCurrentTabItem] = useState('general')
+  const [alertState, setAlertState] = useState({ open: false, content: [] })
 
   const tabItems = [
     { key: 'general', content: t('GENERAL', 'General') },
@@ -29,40 +34,70 @@ const DriversCompanyDetailsFormUI = (props) => {
     { key: 'webhooks', content: t('WEBHOOKS', 'Webhooks') }
   ]
 
+  const closeAlert = () => {
+    cleanActionState()
+    setAlertState({
+      open: false,
+      content: []
+    })
+  }
+
+  useEffect(() => {
+    if (!actionState.loading && actionState.error) {
+      setAlertState({
+        open: true,
+        content: actionState.error || [t('ERROR')]
+      })
+    }
+  }, [actionState])
+
   return (
-    <DetailsContainer>
-      <Header>
-        <h1>
-          {
-            driversCompany
-              ? changesState?.name ?? driversCompany?.name ?? ''
-              : t('DRIVER_COMPANY_SETTINGS', 'Driver company settings')
-          }
-        </h1>
-      </Header>
-      <TabsContainer>
-        <DragScroll>
-          {tabItems.map(item => (
-            <Tab
-              key={item.key}
-              active={item.key === currentTabItem}
-              onClick={() => setCurrentTabItem(item.key)}
-            >
-              {item.content}
-            </Tab>
-          ))}
-        </DragScroll>
-      </TabsContainer>
-      {currentTabItem === 'general' && (
-        <DriversCompanyGeneralDetails {...props} />
-      )}
-      {currentTabItem === 'schedule' && (
-        <DriversCompanyScheduleDetails {...props} />
-      )}
-      {currentTabItem === 'webhooks' && (
-        <DriversCompanyWebhooksDetails {...props} />
-      )}
-    </DetailsContainer>
+    <>
+      <DetailsContainer>
+        <Header>
+          <h1>
+            {
+              driversCompany
+                ? changesState?.name ?? driversCompany?.name ?? ''
+                : t('DRIVER_COMPANY_SETTINGS', 'Driver company settings')
+            }
+          </h1>
+        </Header>
+        <TabsContainer>
+          <DragScroll>
+            {tabItems.map(item => (
+              <Tab
+                key={item.key}
+                active={item.key === currentTabItem}
+                onClick={() => setCurrentTabItem(item.key)}
+              >
+                {item.content}
+              </Tab>
+            ))}
+          </DragScroll>
+        </TabsContainer>
+
+        {currentTabItem === 'general' && (
+          <DriversCompanyGeneralDetails {...props} />
+        )}
+        {currentTabItem === 'schedule' && (
+          <DriversCompanyScheduleDetails {...props} />
+        )}
+        {currentTabItem === 'webhooks' && (
+          <DriversCompanyWebhooksDetails {...props} />
+        )}
+
+      </DetailsContainer>
+      <Alert
+        title={t('ERROR')}
+        content={alertState.content}
+        acceptText={t('ACCEPT')}
+        open={alertState.open}
+        onClose={() => closeAlert()}
+        onAccept={() => closeAlert()}
+        closeOnBackdrop={false}
+      />
+    </>
   )
 }
